Reorganized the tests, so it will get detected by Surefire
