+++ src/main/java/com.ansill.arrays/WriteOnlyByteArray.java
@@ -49,7 +49,6 @@ public interface WriteOnlyByteArray extends ByteArray{
    * @throws ByteArrayInvalidLengthException    thrown if the length is negative
    * @throws ByteArrayLengthOverBoundsException thrown if length goes over the bounds
    */
-  @SuppressWarnings("unchecked")
   @Nonnull
   @Override
   WriteOnlyByteArray subsetOf(@Nonnegative long start, @Nonnegative long length)
