+++ src/test/java/com/ansill/arrays/ReadOnlyPrimitiveByteArrayWithControlByteArrayTest.java
@@ -1,0 +1,11 @@
+package com.ansill.arrays;
+
+import org.junit.jupiter.api.DisplayName;
+import test.ReadOnlyByteArrayWithOtherByteArrayTest;
+
+@DisplayName("PrimitiveByteArray - ReadOnly test with control ByteArray implementation")
+public
+class ReadOnlyPrimitiveByteArrayWithControlByteArrayTest extends ReadOnlyPrimitiveByteArrayTest implements
+  ReadOnlyByteArrayWithOtherByteArrayTest{
+
+}
