+++ src/test/java/com/ansill/arrays/WriteOnlyByteBufferByteArrayWithControlByteArrayTest.java
@@ -1,0 +1,11 @@
+package com.ansill.arrays;
+
+import org.junit.jupiter.api.DisplayName;
+import test.WriteOnlyByteArrayWithOtherByteArrayTest;
+
+@DisplayName("ByteBufferByteArray - WriteOnly test with control ByteArray implementation")
+public
+class WriteOnlyByteBufferByteArrayWithControlByteArrayTest extends WriteOnlyByteBufferByteArrayTest implements
+  WriteOnlyByteArrayWithOtherByteArrayTest{
+
+}
