+++ src/test/java/com/ansill/arrays/WriteOnlyPrimitiveByteArrayWithControlByteArrayTest.java
@@ -1,0 +1,11 @@
+package com.ansill.arrays;
+
+import org.junit.jupiter.api.DisplayName;
+import test.WriteOnlyByteArrayWithOtherByteArrayTest;
+
+@DisplayName("PrimitiveByteArray - WriteOnly test with control ByteArray implementation")
+public
+class WriteOnlyPrimitiveByteArrayWithControlByteArrayTest extends WriteOnlyPrimitiveByteArrayTest implements
+  WriteOnlyByteArrayWithOtherByteArrayTest{
+
+}
