+++ src/test/java/test/ByteArrayTest.java
@@ -17,11 +17,11 @@ import java.util.List;
 import java.util.Random;
 import java.util.Set;
 
+import static com.ansill.arrays.TestUtility.f;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
-import static test.TestUtility.f;
 
 public interface ByteArrayTest{
